+++ blend/crypto/Cargo.toml
@@ -17,9 +17,11 @@ blake2         = { workspace = true }
 lb-groth16     = { workspace = true }
 lb-poq         = { workspace = true }
 lb-poseidon2   = { workspace = true }
-lb-utils       = { features = ["rng"], workspace = true }
+rand           = { features = ["getrandom"], workspace = true }
+rand_chacha    = { workspace = true }
 rs-merkle-tree = { features = ["memory_store"], workspace = true }
 thiserror      = { workspace = true }
 
 [dev-dependencies]
+nistrs     = { workspace = true }
 num-bigint = { workspace = true }
