+++ blend/message/Cargo.toml
@@ -25,7 +25,9 @@ lb-cryptarchia-engine         = { workspace = true }
 lb-groth16                    = { workspace = true }
 lb-key-management-system-keys = { workspace = true }
 lb-log-targets                = { workspace = true }
-lb-utils                      = { features = ["rng"], workspace = true }
+lb-utils                      = { workspace = true }
+rand                          = { features = ["getrandom"], workspace = true }
+rand_chacha                   = { workspace = true }
 serde                         = { workspace = true }
 serde_with                    = { features = ["alloc"], workspace = true }
 thiserror                     = { workspace = true }
