+++ kms/keys/Cargo.toml
@@ -23,7 +23,7 @@ lb-groth16                      = { workspace = true }
 lb-key-management-system-macros = { workspace = true }
 lb-log-targets                  = { workspace = true }
 lb-poseidon2                    = { workspace = true }
-lb-utils                        = { features = ["rng"], workspace = true }
+lb-utils                        = { workspace = true }
 lb-zksign                       = { workspace = true }
 num-bigint                      = { workspace = true }
 rand_core                       = { workspace = true }
