+++ utils/Cargo.toml
@@ -20,8 +20,6 @@ humantime      = { optional = true, workspace = true }
 lb-log-targets = { workspace = true }
 multiaddr      = { workspace = true }
 overwatch      = { workspace = true }
-rand           = { features = ["getrandom"], workspace = true }
-rand_chacha    = { optional = true, workspace = true }
 serde          = { features = ["alloc", "derive"], workspace = true }
 serde_ignored  = { workspace = true }
 serde_with     = { optional = true, workspace = true }
@@ -32,7 +30,6 @@ tokio          = { optional = true, workspace = true }
 tracing        = { workspace = true }
 
 [features]
-rng              = ["dep:rand_chacha"]
 serde            = ["const-hex/alloc", "serde/alloc"]
 time             = ["dep:humantime", "dep:serde_with", "dep:time"]
 tokio            = ["dep:futures", "dep:tokio"]
@@ -40,8 +37,6 @@ tokio-task-names = ["tokio", "tokio/rt", "tokio/tracing"]
 
 [dev-dependencies]
 bincode     = { workspace = true }
-blake2      = { workspace = true }
-nistrs      = { workspace = true }
 serde_json  = { features = ["alloc"], workspace = true }
 serde_with  = { features = ["macros"], workspace = true }
 tempfile    = { workspace = true }
