+++ utils/src/lib.rs
@@ -5,9 +5,6 @@ pub mod noop_service;
 pub mod types;
 pub mod yaml;
 
-#[cfg(feature = "rng")]
-pub mod chacha_rng;
-
 #[cfg(feature = "time")]
 pub mod bounded_duration;
